Decreased arm up and down spped to 0.3, fixed some things in robotContrained, and fixed an argument in AlgaeShooterCommands, and fixed callibrations to robotContainer

+++ src/main/java/frc/robot/commands/AlgaeShooterCommands.java
@@ -93,7 +93,7 @@ public Command getDepositCommand(int depositPosition){
     );
 }*/
 
-public Command getIntakeCommand(int armPosition){
+public Command getIntakeCommand(Double armPosition){
     return Commands.sequence(
         m_armSubsystem.setArmPositionCommand(armPosition),
         Commands.waitUntil(m_atArmPosition),
